requirements.txt
Babel==2.12.1
Bootstrap-Flask==2.2.0
Flask==2.2.3
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.2
PySocks==1.7.1
PyYAML==6.0
SQLAlchemy==2.0.9
WTForms==3.0.1
Werkzeug==2.2.3
ansicolors==1.1.8
async-generator	1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dominate==2.7.0
exceptiongroup==1.1.1
fields-validators==0.1.5
flask-bootstrap5==0.1.dev1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
jinja2-mjml==0.1.0
jinja2-tools==1.0.7
mjml-python==1.0.0
numpy==1.24.2
outcome==1.2.0
packaging==23.1
pandas==2.0.0
pip==23.0.1
psycopg2==2.9.6
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
selenium==4.8.3
setuptools==67.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
validators==0.20.0
visitor==0.1.3
webdriver-manager==3.8.6

